fixed mistakes

+++ src/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: jeperez- <jeperez-@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/10 19:18:46 by jeperez-          #+#    #+#             */
-/*   Updated: 2025/03/05 13:55:07 by jeperez-         ###   ########.fr       */
+/*   Updated: 2025/04/23 14:35:23 by jeperez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				index;
+	size_t	index;
 
 	index = 0;
 	while (index < n)
